refactor: removed duplicated endpoints, optimized friends and reviews services
