+++ src/books/books.service.ts
@@ -399,59 +399,4 @@ export class BooksService {
     return this.booksFileService.sendBookDownloadToResponse(res, book, userId);
   }
 
-  getBookReviewsForAuthor(bookId: string, query: PaginateQuery) {
-    return this.listReviewsForBook(bookId, query);
-  }
-
-  getMyBookReview(readerId: string, bookId: string, query: PaginateQuery) {
-    return this.listReaderReviewForBook(readerId, bookId, query);
-  }
-
-  private listReviewsForBook(bookId: string, query: PaginateQuery) {
-    const qb = this.reviewRepo
-      .createQueryBuilder('review')
-      .leftJoinAndSelect('review.application', 'application')
-      .leftJoinAndSelect('application.reader', 'reader')
-      .leftJoinAndSelect('application.book', 'book')
-      .where('application.bookId = :bookId', { bookId });
-
-    return paginate(query, qb, {
-      sortableColumns: ['createdAt'],
-      defaultSortBy: [['createdAt', 'DESC']],
-      defaultLimit: 20,
-      maxLimit: 100,
-    });
-  }
-
-  private async listReaderReviewForBook(
-    userId: string,
-    bookId: string,
-    query: PaginateQuery,
-  ) {
-    const review = await this.reviewRepo.findOne({
-      where: {
-        application: {
-          bookId,
-          readerId: userId,
-        },
-      },
-      relations: ['application', 'application.reader', 'application.book'],
-    });
-
-    const qb = this.reviewRepo
-      .createQueryBuilder('review')
-      .leftJoinAndSelect('review.application', 'application')
-      .leftJoinAndSelect('application.reader', 'reader')
-      .leftJoinAndSelect('application.book', 'book')
-      .where(review ? 'review.id = :reviewId' : '1 = 0', {
-        reviewId: review?.id,
-      });
-
-    return paginate(query, qb, {
-      sortableColumns: ['createdAt'],
-      defaultSortBy: [['createdAt', 'DESC']],
-      defaultLimit: 20,
-      maxLimit: 100,
-    });
-  }
 }
