+++ src/friends/friends.module.ts
@@ -4,7 +4,12 @@ import { Friend } from './entity/friend.entity';
 import { User } from '../users/entity/user.entity';
 import { FriendsService } from './friends.service';
 import { FriendsController } from './friends.controller';
-import { UserProfileModule } from '../user-profile/user-profile.module';
+import {
+  FriendsListHelper,
+  FriendsNotificationsHelper,
+  FriendsQueryHelper,
+  FriendsSearchHelper,
+} from './helpers';
 import { UserActivityModule } from '../user-activity/user-activity.module';
 import { NotificationModule } from '../notifications/notification.module';
 import { AuthModule } from '../auth/auth.module';
@@ -12,12 +17,17 @@ import { AuthModule } from '../auth/auth.module';
 @Module({
   imports: [
     TypeOrmModule.forFeature([Friend, User]),
-    forwardRef(() => UserProfileModule),
     forwardRef(() => UserActivityModule),
     forwardRef(() => NotificationModule),
     AuthModule,
   ],
-  providers: [FriendsService],
+  providers: [
+    FriendsService,
+    FriendsQueryHelper,
+    FriendsListHelper,
+    FriendsSearchHelper,
+    FriendsNotificationsHelper,
+  ],
   controllers: [FriendsController],
   exports: [FriendsService],
 })
