+++ src/reviews/dto/find-reviews.dto.ts
@@ -4,7 +4,10 @@ import { ApiPropertyOptional } from '@nestjs/swagger';
 import { BasePaginationDto } from '../../common/dto/base-pagination.dto';
 
 export class FindReviewsDto extends BasePaginationDto {
-  @ApiPropertyOptional()
+  @ApiPropertyOptional({
+    description:
+      'Only used by GET /reviews/users/:userId. Book reviews use role-based visibility instead.',
+  })
   @IsOptional()
   @Type(() => Boolean)
   @IsBoolean()
