+++ src/user-profile/user-profile.controller.ts
@@ -1,17 +1,10 @@
 import {
-  Body,
   Controller,
-  Delete,
   Get,
   Param,
-  ParseUUIDPipe,
-  Patch,
-  Post,
   Query,
   Request,
   UseGuards,
-  UsePipes,
-  ValidationPipe,
 } from '@nestjs/common';
 import {
   ApiTags,
@@ -21,15 +14,7 @@ import {
 } from '@nestjs/swagger';
 import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
 import { UserProfileService } from './user-profile.service';
-import { UserActivityService } from '../user-activity/user-activity.service';
-import { UserAddressService } from '../user-address/user-address.service';
 import { getUserId } from '../common';
-import { CreateAddressDto, UpdateAddressDto } from '../user-address/dto';
-import {
-  UpdateNotificationSettingsDto,
-  UpdatePrivacySettingsDto,
-  UpdateSocialMediaDto,
-} from './dto';
 import { Request as ExpressRequest } from 'express';
 
 @ApiTags('User Profiles')
@@ -37,108 +22,20 @@ import { Request as ExpressRequest } from 'express';
 @UseGuards(JwtAuthGuard)
 @ApiBearerAuth()
 export class UserProfileController {
-  constructor(
-    private readonly userProfileService: UserProfileService,
-    private readonly userActivityService: UserActivityService,
-    private readonly userAddressService: UserAddressService,
-  ) {}
-
-  @Post('me/addresses')
-  @ApiOperation({
-    summary: 'Create a new address (Authenticated)',
-  })
-  @ApiResponse({ status: 201, description: 'Address created successfully' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  @UsePipes(new ValidationPipe({ transform: true, whitelist: true }))
-  async createAddress(
-    @Request() req: ExpressRequest,
-    @Body() dto: CreateAddressDto,
-  ) {
-    const userId = getUserId(req);
-    return this.userAddressService.create(userId, dto);
-  }
+  constructor(private readonly userProfileService: UserProfileService) {}
 
   @Get('user/:usernameOrId')
   @ApiOperation({
     summary: 'Get public profile by username or user ID (Authenticated)',
   })
   @ApiResponse({ status: 200, description: 'Public user profile' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
   @ApiResponse({ status: 404, description: 'User not found' })
-  async getPublicProfile(
+  getPublicProfile(
     @Request() req: ExpressRequest,
     @Param('usernameOrId') usernameOrId: string,
   ) {
     const viewerId = getUserId(req);
     return this.userProfileService.getPublicProfile(usernameOrId, viewerId);
-  }
-
-  @Get('me')
-  @ApiOperation({ summary: 'Get current user profile (Authenticated)' })
-  @ApiResponse({ status: 200, description: 'User profile' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  async getMyProfile(@Request() req: ExpressRequest) {
-    const userId = getUserId(req);
-    return this.userProfileService.getProfile(userId);
-  }
-
-  @Get('me/activity')
-  @ApiOperation({ summary: 'Get current user activity (Authenticated)' })
-  @ApiResponse({ status: 200, description: 'User activity' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  async getMyActivity(
-    @Request() req: ExpressRequest,
-    @Param('limit') limit?: number,
-  ) {
-    const userId = getUserId(req);
-    return this.userActivityService.getUserActivity(userId, limit);
-  }
-
-  @Get('me/activity/public')
-  @ApiOperation({ summary: 'Get current user public activity (Authenticated)' })
-  @ApiResponse({ status: 200, description: 'Public user activity' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  async getMyPublicActivity(
-    @Request() req: ExpressRequest,
-    @Param('limit') limit?: number,
-  ) {
-    const userId = getUserId(req);
-    return this.userActivityService.getPublicActivity(userId, limit);
-  }
-
-  @Get('me/activity/recent')
-  @ApiOperation({ summary: 'Get current user recent activity (Authenticated)' })
-  @ApiResponse({ status: 200, description: 'Recent user activity' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  async getMyRecentActivity(
-    @Request() req: ExpressRequest,
-    @Param('days') days?: number,
-    @Param('limit') limit?: number,
-  ) {
-    const userId = getUserId(req);
-    return this.userActivityService.getRecentActivity(userId, days, limit);
-  }
-
-  @Get('me/activity/stats')
-  @ApiOperation({
-    summary: 'Get current user activity statistics (Authenticated)',
-  })
-  @ApiResponse({ status: 200, description: 'User activity statistics' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  async getMyActivityStats(@Request() req: ExpressRequest) {
-    const userId = getUserId(req);
-    return this.userActivityService.getActivityStats(userId);
-  }
-
-  @Get('me/addresses')
-  @ApiOperation({
-    summary: 'Get current user addresses (Authenticated)',
-  })
-  @ApiResponse({ status: 200, description: 'List of user addresses' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  async getMyAddresses(@Request() req: ExpressRequest) {
-    const userId = getUserId(req);
-    return this.userAddressService.findByUserId(userId);
   }
 
   @Get('user/:usernameOrId/activity/recent')
@@ -147,10 +44,9 @@ export class UserProfileController {
       'Get user public recent activity by username or ID (Authenticated)',
   })
   @ApiResponse({ status: 200, description: 'User recent public activity' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
   @ApiResponse({ status: 403, description: 'Activity is private' })
   @ApiResponse({ status: 404, description: 'User not found' })
-  async getUserRecentActivity(
+  getUserRecentActivity(
     @Request() req: ExpressRequest,
     @Param('usernameOrId') usernameOrId: string,
     @Query('days') days?: number,
@@ -163,104 +59,5 @@ export class UserProfileController {
       days ? parseInt(days.toString(), 10) : 7,
       limit ? parseInt(limit.toString(), 10) : 50,
     );
-  }
-
-  @Patch('me/addresses/:addressId')
-  @ApiOperation({
-    summary: 'Update an address (Authenticated)',
-  })
-  @ApiResponse({ status: 200, description: 'Address updated successfully' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  @ApiResponse({ status: 404, description: 'Address not found' })
-  @UsePipes(new ValidationPipe({ transform: true, whitelist: true }))
-  async updateAddress(
-    @Request() req: ExpressRequest,
-    @Param('addressId', ParseUUIDPipe) addressId: string,
-    @Body() dto: UpdateAddressDto,
-  ) {
-    const userId = getUserId(req);
-    return this.userAddressService.update(userId, addressId, dto);
-  }
-
-  @Patch('me/social-media')
-  @ApiOperation({
-    summary: 'Update social media links (Authenticated)',
-  })
-  @ApiResponse({
-    status: 200,
-    description: 'Social media updated successfully',
-  })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  @UsePipes(new ValidationPipe({ transform: true, whitelist: true }))
-  async updateSocialMedia(
-    @Request() req: ExpressRequest,
-    @Body() dto: UpdateSocialMediaDto,
-  ) {
-    const userId = getUserId(req);
-
-    const socialMedia = {
-      instagram: dto.instagram,
-      tiktok: dto.tiktok,
-      youtube: dto.youtube,
-      goodreads: dto.goodreads,
-      custom: dto.custom,
-    };
-    return this.userProfileService.updateSocialMedia(userId, socialMedia);
-  }
-
-  @Patch('me/privacy')
-  @ApiOperation({
-    summary: 'Update privacy settings (Authenticated)',
-  })
-  @ApiResponse({
-    status: 200,
-    description: 'Privacy settings updated successfully',
-  })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  @UsePipes(new ValidationPipe({ transform: true, whitelist: true }))
-  async updatePrivacySettings(
-    @Request() req: ExpressRequest,
-    @Body() dto: UpdatePrivacySettingsDto,
-  ) {
-    const userId = getUserId(req);
-    return this.userProfileService.updatePrivacySettings(userId, dto);
-  }
-
-  @Patch('me/notifications')
-  @ApiOperation({
-    summary: 'Update notification settings (Authenticated)',
-  })
-  @ApiResponse({
-    status: 200,
-    description: 'Notification settings updated successfully',
-  })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  @UsePipes(new ValidationPipe({ transform: true, whitelist: true }))
-  async updateNotificationSettings(
-    @Request() req: ExpressRequest,
-    @Body() dto: UpdateNotificationSettingsDto,
-  ) {
-    const userId = getUserId(req);
-    return this.userProfileService.updateNotificationSettings(userId, {
-      notificationsEnabled: dto.notificationsEnabled,
-      emailNotifications: dto.emailNotifications,
-      notificationPreferences: dto.notificationPreferences,
-    });
-  }
-
-  @Delete('me/addresses/:addressId')
-  @ApiOperation({
-    summary: 'Delete an address (Authenticated)',
-  })
-  @ApiResponse({ status: 200, description: 'Address deleted successfully' })
-  @ApiResponse({ status: 401, description: 'Unauthorized' })
-  @ApiResponse({ status: 404, description: 'Address not found' })
-  async deleteAddress(
-    @Request() req: ExpressRequest,
-    @Param('addressId', ParseUUIDPipe) addressId: string,
-  ) {
-    const userId = getUserId(req);
-    await this.userAddressService.remove(userId, addressId);
-    return { message: 'Address deleted successfully' };
   }
 }
