+++ src/user-profile/user-profile.module.ts
@@ -4,6 +4,7 @@ import { UserProfile } from './entity/user-profile.entity';
 import { User } from '../users/entity/user.entity';
 import { UserProfileService } from './user-profile.service';
 import { UserProfileController } from './user-profile.controller';
+import { UserProfileMeController } from './user-profile-me.controller';
 import { FriendsModule } from '../friends/friends.module';
 import { UserActivityModule } from '../user-activity/user-activity.module';
 import { UserAddressModule } from '../user-address/user-address.module';
@@ -20,7 +21,7 @@ import { UsersModule } from '../users/users.module';
     UserAddressModule,
   ],
   providers: [UserProfileService],
-  controllers: [UserProfileController],
+  controllers: [UserProfileController, UserProfileMeController],
   exports: [UserProfileService],
 })
 export class UserProfileModule {}
