+++ src/users/users.controller.ts
@@ -7,6 +7,7 @@ import {
   Patch,
   Post,
   Delete,
+  Query,
   UseGuards,
   UseInterceptors,
   UploadedFile,
@@ -21,12 +22,17 @@ import {
   ApiBearerAuth,
   ApiConsumes,
   ApiBody,
+  ApiQuery,
 } from '@nestjs/swagger';
 import { memoryStorage } from 'multer';
 import { UsersService } from './users.service';
 import { UpdateProfileDto, UploadAvatarDto } from './dto';
 import { UserResponseDto, UserPublicResponseDto } from './dto';
-import { JwtAuthGuard } from '../auth/guards';
+import { JwtAuthGuard, RolesGuard } from '../auth/guards';
+import { Roles } from '../auth/decorators';
+import { UserType } from './enums';
+import { ReviewsService } from '../reviews/reviews.service';
+import { Review } from '../reviews/entity/review.entity';
 import {
   CurrentUser,
   JwtPayload,
@@ -35,7 +41,10 @@ import {
 @ApiTags('Users')
 @Controller('users')
 export class UsersController {
-  constructor(private readonly usersService: UsersService) {}
+  constructor(
+    private readonly usersService: UsersService,
+    private readonly reviewsService: ReviewsService,
+  ) {}
   @UseGuards(JwtAuthGuard)
   @Post('me/avatar')
   @ApiBearerAuth()
@@ -78,6 +87,34 @@ export class UsersController {
   @ApiResponse({ status: 200, description: 'User statistics' })
   async getMyStats(@CurrentUser('sub') userId: string) {
     return this.usersService.getMyStats(userId);
+  }
+
+  @UseGuards(JwtAuthGuard, RolesGuard)
+  @Roles(UserType.AUTHOR)
+  @Get('me/reviews/latest')
+  @ApiBearerAuth()
+  @ApiOperation({
+    summary: 'Get latest reviews across all books (Author only)',
+  })
+  @ApiQuery({
+    name: 'limit',
+    required: false,
+    type: Number,
+    description: 'Number of reviews to return (default: 3)',
+  })
+  @ApiResponse({
+    status: 200,
+    description: 'List of latest reviews across all author books',
+    type: [Review],
+  })
+  getMyLatestReviews(
+    @CurrentUser('sub') authorId: string,
+    @Query('limit') limit?: number,
+  ) {
+    return this.reviewsService.getAuthorLatestReviews(
+      authorId,
+      limit ? parseInt(limit.toString(), 10) : 3,
+    );
   }
 
   @UseGuards(JwtAuthGuard)
