+++ src/users/users.module.ts
@@ -8,6 +8,7 @@ import { Review } from '../reviews/entity/review.entity';
 import { AuthModule } from '../auth/auth.module';
 import { FilesModule } from '../files/files.module';
 import { BooksModule } from '../books/books.module';
+import { ReviewsModule } from '../reviews/reviews.module';
 import { UsersReaderStatsHelper } from './helpers/users-reader-stats.helper';
 
 @Module({
@@ -16,6 +17,7 @@ import { UsersReaderStatsHelper } from './helpers/users-reader-stats.helper';
     AuthModule,
     FilesModule,
     BooksModule,
+    ReviewsModule,
   ],
   controllers: [UsersController],
   providers: [UsersService, UsersReaderStatsHelper],
